refactor: refactoring the middleware.test.js into their repsctive modules
