+++ test/middleware.test.js
@@ -1,22 +1,11 @@
-import fs from 'fs';
-import path from 'path';
-
 import express from 'express';
 import request from 'supertest';
-import del from 'del';
 
 import middleware from '../src';
 
 import getCompiler from './helpers/getCompiler';
-import GetLogsPlugin from './helpers/GetLogsPlugin';
 
 import webpackConfig from './fixtures/webpack.config';
-import webpackSimpleConfig from './fixtures/webpack.simple.config';
-import webpackMultiConfig from './fixtures/webpack.array.config';
-import webpackErrorConfig from './fixtures/webpack.error.config';
-import webpackMultiErrorConfig from './fixtures/webpack.array.error.config';
-import webpackWarningConfig from './fixtures/webpack.warning.config';
-import webpackMultiWarningConfig from './fixtures/webpack.array.warning.config';
 
 describe('middleware', () => {
   let instance;
@@ -195,682 +184,6 @@ describe('middleware', () => {
 
           return done();
         });
-    });
-  });
-
-  describe('index option', () => {
-    describe('should work with "false" value', () => {
-      beforeAll((done) => {
-        const compiler = getCompiler(webpackConfig);
-
-        instance = middleware(compiler, { index: false, publicPath: '/' });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should return the "404" code for the "GET" request to the public path', (done) => {
-        request(app)
-          .get('/')
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(404, done);
-      });
-
-      it('should return the "200" code for the "GET" request to the "index.html" file', (done) => {
-        request(app)
-          .get('/index.html')
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(200, done);
-      });
-    });
-
-    describe('should work with "true" value', () => {
-      beforeAll((done) => {
-        const compiler = getCompiler(webpackConfig);
-
-        instance = middleware(compiler, { index: true, publicPath: '/' });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request to the public path', (done) => {
-        request(app)
-          .get('/')
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(200, done);
-      });
-
-      it('should return the "200" code for the "GET" request to the public path', (done) => {
-        request(app)
-          .get('/index.html')
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(200, done);
-      });
-    });
-
-    describe('should work with "string" value', () => {
-      beforeAll((done) => {
-        const outputPath = path.resolve(__dirname, './outputs/basic');
-        const compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: outputPath,
-          },
-        });
-
-        instance = middleware(compiler, {
-          index: 'default.html',
-          publicPath: '/',
-        });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-
-        instance.context.outputFileSystem.mkdirSync(outputPath, {
-          recursive: true,
-        });
-        instance.context.outputFileSystem.writeFileSync(
-          path.resolve(outputPath, 'default.html'),
-          'hello'
-        );
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request to the public path', (done) => {
-        request(app)
-          .get('/')
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(200, done);
-      });
-    });
-
-    describe('should work with "string" value with a custom extension', () => {
-      beforeAll((done) => {
-        const outputPath = path.resolve(__dirname, './outputs/basic');
-        const compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: outputPath,
-          },
-        });
-
-        instance = middleware(compiler, {
-          index: 'index.custom',
-          publicPath: '/',
-        });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-
-        instance.context.outputFileSystem.mkdirSync(outputPath, {
-          recursive: true,
-        });
-        instance.context.outputFileSystem.writeFileSync(
-          path.resolve(outputPath, 'index.custom'),
-          'hello'
-        );
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request to the public path', (done) => {
-        request(app)
-          .get('/')
-          .expect('Content-Type', 'application/octet-stream')
-          .expect(200, done);
-      });
-    });
-
-    describe('should work with "string" value with a custom extension and defined a custom MIME type', () => {
-      beforeAll((done) => {
-        const outputPath = path.resolve(__dirname, './outputs/basic');
-        const compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: outputPath,
-          },
-        });
-
-        instance = middleware(compiler, {
-          index: 'index.custom',
-          mimeTypes: {
-            custom: 'text/html',
-          },
-          publicPath: '/',
-        });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-
-        instance.context.outputFileSystem.mkdirSync(outputPath, {
-          recursive: true,
-        });
-        instance.context.outputFileSystem.writeFileSync(
-          path.resolve(outputPath, 'index.custom'),
-          'hello'
-        );
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request to the public path', (done) => {
-        request(app)
-          .get('/')
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(200, done);
-      });
-    });
-
-    describe('should work with "string" value without an extension', () => {
-      beforeAll((done) => {
-        const outputPath = path.resolve(__dirname, './outputs/basic');
-        const compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: outputPath,
-          },
-        });
-
-        instance = middleware(compiler, { index: 'noextension' });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-
-        instance.context.outputFileSystem.mkdirSync(outputPath, {
-          recursive: true,
-        });
-        instance.context.outputFileSystem.writeFileSync(
-          path.resolve(outputPath, 'noextension'),
-          'hello'
-        );
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request to the public path', (done) => {
-        request(app)
-          .get('/')
-          .expect('Content-Type', 'application/octet-stream')
-          .expect(200, done);
-      });
-    });
-
-    describe('should work with "string" value but the "index" option is a directory', () => {
-      beforeAll((done) => {
-        const outputPath = path.resolve(__dirname, './outputs/basic');
-        const compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: outputPath,
-          },
-        });
-
-        instance = middleware(compiler, {
-          index: 'custom.html',
-          publicPath: '/',
-        });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-
-        instance.context.outputFileSystem.mkdirSync(outputPath, {
-          recursive: true,
-        });
-        instance.context.outputFileSystem.mkdirSync(
-          path.resolve(outputPath, 'custom.html')
-        );
-      });
-
-      afterAll(close);
-
-      it('should return the "404" code for the "GET" request to the public path', (done) => {
-        request(app).get('/').expect(404, done);
-      });
-    });
-
-    describe('should not handle request when index is neither a file nor a directory', () => {
-      let compiler;
-      let isDirectory;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackConfig);
-
-        instance = middleware(compiler, {
-          index: 'default.html',
-          publicPath: '/',
-        });
-
-        isDirectory = jest
-          .spyOn(instance.context.outputFileSystem, 'statSync')
-          .mockImplementation(() => {
-            return {
-              isFile: () => false,
-              isDirectory: () => false,
-            };
-          });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll((done) => {
-        isDirectory.mockRestore();
-
-        close(done);
-      });
-
-      it('should return the "404" code for the "GET" request to the public path', (done) => {
-        request(app).get('/').expect(404, done);
-      });
-    });
-  });
-
-  describe('logger', () => {
-    describe('should logging on successfully build', () => {
-      let compiler;
-      let getLogsPlugin;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackConfig);
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should logging', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            instance.invalidate();
-
-            return instance.waitUntilValid(() => {
-              expect(getLogsPlugin.logs).toMatchSnapshot();
-
-              done();
-            });
-          });
-      });
-    });
-
-    describe('should logging on successfully build in multi-compiler mode', () => {
-      let compiler;
-      let getLogsPlugin;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackMultiConfig);
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should logging', (done) => {
-        request(app)
-          .get('/static-one/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            instance.invalidate();
-
-            return instance.waitUntilValid(() => {
-              expect(getLogsPlugin.logs).toMatchSnapshot();
-
-              done();
-            });
-          });
-      });
-    });
-
-    describe('should logging on unsuccessful build', () => {
-      let compiler;
-      let getLogsPlugin;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackErrorConfig);
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should logging', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            instance.invalidate();
-
-            return instance.waitUntilValid(() => {
-              expect(getLogsPlugin.logs).toMatchSnapshot();
-
-              done();
-            });
-          });
-      });
-    });
-
-    describe('should logging on unsuccessful build in multi-compiler ', () => {
-      let compiler;
-      let getLogsPlugin;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackMultiErrorConfig);
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should logging', (done) => {
-        request(app)
-          .get('/static-one/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            instance.invalidate();
-
-            return instance.waitUntilValid(() => {
-              expect(getLogsPlugin.logs).toMatchSnapshot();
-
-              done();
-            });
-          });
-      });
-    });
-
-    describe('should logging an warning', () => {
-      let compiler;
-      let getLogsPlugin;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackWarningConfig);
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should logging', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            instance.invalidate();
-
-            return instance.waitUntilValid(() => {
-              expect(getLogsPlugin.logs).toMatchSnapshot();
-
-              done();
-            });
-          });
-      });
-    });
-
-    describe('should logging warnings in multi-compiler mode', () => {
-      let compiler;
-      let getLogsPlugin;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackMultiWarningConfig);
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should logging', (done) => {
-        request(app)
-          .get('/static-one/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            instance.invalidate();
-
-            return instance.waitUntilValid(() => {
-              expect(getLogsPlugin.logs).toMatchSnapshot();
-
-              done();
-            });
-          });
-      });
-    });
-
-    describe('should logging an error in "watch" method', () => {
-      let getLogsPlugin;
-
-      it('should logging on startup', () => {
-        const compiler = getCompiler(webpackConfig);
-
-        const watchSpy = jest
-          .spyOn(compiler, 'watch')
-          .mockImplementation((watchOptions, callback) => {
-            const error = new Error('Error in Watch method');
-
-            error.stack = '';
-
-            callback(error);
-
-            return { close: () => {} };
-          });
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        expect(getLogsPlugin.logs).toMatchSnapshot();
-
-        instance.close();
-
-        watchSpy.mockRestore();
-      });
-    });
-
-    describe('should logging an error from the "fs.mkdir" method when the "writeToDisk" option is "true" ', () => {
-      let compiler;
-      let getLogsPlugin;
-      let mkdirSpy;
-
-      beforeAll((done) => {
-        compiler = getCompiler({
-          ...webpackSimpleConfig,
-          output: {
-            filename: 'bundle.js',
-            path: path.resolve(
-              __dirname,
-              './outputs/write-to-disk-mkdir-error'
-            ),
-          },
-        });
-
-        mkdirSpy = jest.spyOn(fs, 'mkdir').mockImplementation((...args) => {
-          const callback = args[args.length - 1];
-
-          return callback(new Error('Error in the "fs.mkdir" method.'));
-        });
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler, { writeToDisk: true });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll((done) => {
-        del.sync(
-          path.posix.resolve(__dirname, './outputs/write-to-disk-mkdir-error')
-        );
-
-        mkdirSpy.mockRestore();
-
-        close(done);
-      });
-
-      it('should logging', (done) => {
-        compiler.hooks.failed.tap('FailedCatcher', () => {
-          instance.close(() => {
-            expect(getLogsPlugin.logs).toMatchSnapshot();
-
-            done();
-          });
-        });
-      });
-    });
-
-    describe('should logging an error from the "fs.writeFile" method when the "writeToDisk" option is "true" ', () => {
-      let compiler;
-      let getLogsPlugin;
-      let writeFileSpy;
-
-      beforeAll((done) => {
-        compiler = getCompiler({
-          ...webpackSimpleConfig,
-          output: {
-            filename: 'bundle.js',
-            path: path.resolve(
-              __dirname,
-              './outputs/write-to-disk-writeFile-error'
-            ),
-          },
-        });
-
-        writeFileSpy = jest
-          .spyOn(fs, 'writeFile')
-          .mockImplementation((...args) => {
-            const callback = args[args.length - 1];
-
-            return callback(new Error('Error in the "fs.writeFile" method.'));
-          });
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler, { writeToDisk: true });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll((done) => {
-        writeFileSpy.mockRestore();
-
-        del.sync(
-          path.posix.resolve(
-            __dirname,
-            './outputs/write-to-disk-writeFile-error'
-          )
-        );
-
-        close(done);
-      });
-
-      it('should logging', (done) => {
-        compiler.hooks.failed.tap('FailedCatcher', () => {
-          instance.close(() => {
-            expect(getLogsPlugin.logs).toMatchSnapshot();
-
-            done();
-          });
-        });
-      });
     });
   });
 });
